% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{apply_classification_tree}
\alias{apply_classification_tree}
\title{Apply classification tree to \code{klassets_response_xy} object}
\usage{
apply_classification_tree(df, maxdepth = Inf, alpha = 0.05, type = "prob", ...)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{maxdepth}{Max depth of the tree. Same used in \code{partykit::ctree_control}.}

\item{alpha}{Alpha value, same used in \code{partykit::ctree_control}}

\item{type}{Type of prediction, one of "prob, response, node.}

\item{...}{Options for \code{partykit::ctree_control}.}
}
\description{
Apply classification tree to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 1000, relationship = function(x, y) x**2 > sin(y))

plot(df)

# default type = "prob"
df_tree_prob <- apply_classification_tree(df)
df_tree_prob

df_tree_resp <- apply_classification_tree(df, type = "response")
df_tree_resp

df_tree_node <- apply_classification_tree(df, type = "node")
df_tree_node

plot(df_tree_prob)
plot(df_tree_resp)
plot(df_tree_node)

}
