% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{apply_kmeans_clust}
\alias{apply_kmeans_clust}
\title{Apply K-means to \code{klassets_cluster} object}
\usage{
apply_kmeans_clust(df, k = 3, ...)
}
\arguments{
\item{df}{A \code{klassets_cluster} object.}

\item{k}{A numeric value to pass to \code{stats::kmeans} method.}

\item{...}{Extra parameter for \code{stats::kmeans} function.}
}
\description{
Apply K-means to \code{klassets_cluster} object
}
\examples{

set.seed(123456)

df <- sim_clusters()

plot(apply_kmeans_clust(df, k = 4))

df2 <- sim_clusters(groups = 2, add_group_col = TRUE)

plot(apply_kmeans_clust(df2, k = 2))

}
