% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{apply_knn}
\alias{apply_knn}
\title{Apply K Nearest Neighbours to \code{klassets_response_xy} object}
\usage{
apply_knn(df, neighbours = 10, type = "prob")
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{neighbours}{The neighbours parameter.}

\item{type}{Type of prediction, one of prob or response.}
}
\description{
Apply K Nearest Neighbours to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 1000, relationship = function(x, y) x**2 > sin(y))

plot(df)

# defaults to prob
apply_knn(df)

apply_knn(df, type = "response")

plot(apply_knn(df))

plot(apply_knn(df, neighbours = 3))

plot(apply_knn(df, neighbours = 10))

plot(apply_knn(df, neighbours = 200))

plot(apply_knn(df, neighbours = 3, type = "reponse"))

plot(apply_knn(df, neighbours = 10, type = "reponse"))

plot(apply_knn(df, neighbours = 200, type = "reponse"))

}
