% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{apply_linear_model}
\alias{apply_linear_model}
\title{Apply Linear model to \code{klassets_xy} object}
\usage{
apply_linear_model(df, order = 1, stepwise = FALSE, verbose = FALSE)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{order}{Order of predictive variable x.}

\item{stepwise}{A logical value to indicate to perform stepwise.}

\item{verbose}{A logical value to incdicate to show the trace of thet
stepwise procedure.#'}
}
\description{
Apply Linear model to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dflm <- apply_linear_model(df)

dflm

plot(dflm)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 10 * sin(x) + sqrt(abs(x)))

plot(df)

plot(apply_linear_model(df))

plot(apply_linear_model(df, order = 5, stepwise = TRUE, verbose = TRUE))

}
