% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{apply_linear_model_tree}
\alias{apply_linear_model_tree}
\title{Apply Linear Model tree to \code{klassets_xy} object}
\usage{
apply_linear_model_tree(df, maxdepth = Inf, alpha = 0.05, ...)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{...}{Options for \code{partykit::mltree}.}
}
\description{
Apply Linear Model tree to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dflm <- apply_linear_model_tree(df)

dflm

plot(dflm)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 10 * sin(x) + sqrt(abs(x)))

plot(df)

plot(apply_linear_model_tree(df))

}
