% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{apply_statskmeans}
\alias{apply_statskmeans}
\title{Apply K-means to \code{klassets_cluster} object using \code{stats::kmeans}}
\usage{
apply_statskmeans(df, centers = 3, ...)
}
\arguments{
\item{df}{A \code{klassets_cluster} object.}

\item{centers}{A numeric value to pass to \code{stats::kmeans} method.
The famous k parameter.}

\item{...}{Extra parameter for \code{stats::kmeans} function.}
}
\description{
Apply K-means to \code{klassets_cluster} object using \code{stats::kmeans}
}
\examples{

set.seed(12)

df <- sim_groups(n = 200, groups = 3)

plot(df)

set.seed(124)

dfc <- apply_statskmeans(df, centers = 4)

plot(dfc)

}
