% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{fit_classification_random_forest}
\alias{fit_classification_random_forest}
\title{Fit classification random forest to \code{klassets_response_xy} object}
\usage{
fit_classification_random_forest(
  df,
  type = "prob",
  ntree = 500L,
  maxdepth = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{type}{Type of prediction, one of prob, response, node.}

\item{ntree}{Number of trees to grow for the forest.}

\item{maxdepth}{Max depth of each trees.}

\item{trace}{A logical indicating if a progress bar shall be printed while the forest grows.}

\item{...}{Options for \code{ranger::ranger}.}
}
\description{
Fit classification random forest to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 1000, relationship = function(x, y) x**2 > sin(y))

plot(df)

dfcrf <- fit_classification_random_forest(df)

dfcrf

plot(dfcrf)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 3 * sin(x) + 5 * sqrt(abs(x)))

plot(df)

plot(fit_regression_random_forest(df))

# default
plot(fit_regression_random_forest(df))

}
