% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{fit_classification_tree}
\alias{fit_classification_tree}
\title{Fit classification tree to \code{klassets_response_xy} object}
\usage{
fit_classification_tree(df, type = "prob", maxdepth = Inf, alpha = 0.05, ...)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{type}{Type of prediction, one of prob, response, node.}

\item{maxdepth}{Max depth of the tree. Same used in \code{partykit::ctree_control}.}

\item{alpha}{Alpha value, same used in \code{partykit::ctree_control}}

\item{...}{Options for \code{partykit::ctree_control}.}
}
\description{
Fit classification tree to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 1000, relationship = function(x, y) x**2 > sin(y))

plot(df)

# default type = "prob"
df_tree_prob <- fit_classification_tree(df)
df_tree_prob

df_tree_resp <- fit_classification_tree(df, type = "response")
df_tree_resp

df_tree_node <- fit_classification_tree(df, type = "node")
df_tree_node

plot(df_tree_prob)
plot(df_tree_resp)
plot(df_tree_node)

}
