% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{fit_kmeans}
\alias{fit_kmeans}
\title{Fit K-means to \code{klassets_cluster} object}
\usage{
fit_kmeans(df, centers = 3, ...)
}
\arguments{
\item{df}{A \code{klassets_cluster} object. A object from \code{sim_groups}.}

\item{centers}{A numeric value to pass to \code{kmeans_iterations} function
The famous k parameter.}

\item{...}{Extra parameters for \code{kmeans_iterations} function.}
}
\description{
Fit K-means to \code{klassets_cluster} object
}
\examples{

set.seed(12)

df <- sim_groups(n = 200, groups = 3)

plot(df)

set.seed(124)

dfc <- fit_kmeans(df, centers = 4, max_iterations = 6)

plot(dfc)

}
