% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{fit_knn}
\alias{fit_knn}
\title{Fit K Nearest Neighbours to \code{klassets_response_xy} object}
\usage{
fit_knn(df, neighbours = 10, type = "prob")
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{neighbours}{The neighbours parameter.}

\item{type}{Type of prediction, one of prob or response.}
}
\description{
Fit K Nearest Neighbours to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 1000, relationship = function(x, y) x**2 > sin(y))

plot(df)

# defaults to prob
fit_knn(df)

fit_knn(df, type = "response")

plot(fit_knn(df))

plot(fit_knn(df, neighbours = 3))

plot(fit_knn(df, neighbours = 10))

plot(fit_knn(df, neighbours = 200))

plot(fit_knn(df, neighbours = 3, type = "response"))

plot(fit_knn(df, neighbours = 10, type = "response"))

plot(fit_knn(df, neighbours = 200, type = "response"))

}
