% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_linear_model_tree}
\alias{fit_linear_model_tree}
\title{Fit Linear Model tree to \code{klassets_xy} object}
\usage{
fit_linear_model_tree(df, maxdepth = Inf, alpha = 0.05, ...)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{maxdepth}{Max depth of the tree. Same used in \code{partykit::mob_control}.}

\item{alpha}{Alpha value, same used in \code{partykit::mob_control}}

\item{...}{Addiotional options passed to \code{partykit::mob_control}.}
}
\description{
Fit Linear Model tree to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dflm <- fit_linear_model_tree(df)

dflm

plot(dflm)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 10 * sin(x) + sqrt(abs(x)))

plot(df)

plot(fit_linear_model_tree(df))

}
