% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_loess}
\alias{fit_loess}
\title{Fit Local polynomial regression to \code{klassets_xy} object}
\usage{
fit_loess(df, ...)
}
\arguments{
\item{df}{A object from \code{sim_xy}.}

\item{...}{Options for \code{stats::loess}.}
}
\description{
Fit Local polynomial regression to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dfloess <- fit_loess(df)

dfloess

plot(dfloess)

df <- sim_xy(n = 1000, x_dist = runif)
df <- dplyr::mutate(df, y = y + 2*sin(5 * x))
plot(df)

plot(fit_loess(df))

}
