% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_mars}
\alias{fit_mars}
\title{Fit Multivariate Adaptive Regression Splines to \code{klassets_xy} object}
\usage{
fit_mars(df, ...)
}
\arguments{
\item{df}{A object from \code{sim_xy}.}

\item{...}{Options for \code{earth ::earth}.}
}
\description{
Fit Multivariate Adaptive Regression Splines to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dfmars <- fit_mars(df)

dfmars

plot(dfmars)

df <- sim_xy(n = 1000, x_dist = runif)
df <- dplyr::mutate(df, y = y + 2*sin(5 * x))
plot(df)

plot(fit_mars(df))

}
