% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_regression_random_forest}
\alias{fit_regression_random_forest}
\title{Fit regression random forest to \code{klassets_xy} object}
\usage{
fit_regression_random_forest(
  df,
  ntree = 500L,
  maxdepth = Inf,
  trace = FALSE,
  ...
)
}
\arguments{
\item{df}{A object from \code{sim_xy}.}

\item{ntree}{Number of trees to grow for the forest.}

\item{maxdepth}{Max depth of each trees.}

\item{trace}{A logical indicating if a progress bar shall be printed while the forest grows.}

\item{...}{Options for \code{ranger::ranger}.}
}
\description{
Fit regression random forest to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dfrrf <- fit_regression_random_forest(df)

dfrrf

plot(dfrrf)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 3 * sin(x) + 5 * sqrt(abs(x)))

plot(df)

plot(fit_regression_random_forest(df))

}
