% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idyob10k}
\alias{idyob10k}
\title{10,000 observations of ID and year of birth}
\format{
A data frame (klassets class) with 10000 rows and 2 variables.
}
\usage{
idyob10k
}
\description{
A data frame where x is an ID and y is year of birth.
}
\keyword{datasets}
