% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{kmeans_iterations}
\alias{kmeans_iterations}
\title{Generate intermediate iterations when performing K-means}
\usage{
kmeans_iterations(
  df,
  centers = 3,
  tolerance = 1e-05,
  max_iterations = 15,
  verbose = FALSE
)
}
\arguments{
\item{df}{A object from \code{sim_clusters}.}

\item{centers}{How many clusters}

\item{tolerance}{A value to indicating early stop.}

\item{max_iterations}{Max iterations to calculate.}

\item{verbose}{A logical value, to show or not iterations messages.}
}
\description{
Generate intermediate iterations when performing K-means
}
\examples{

set.seed(123456)

df <- sim_clusters(n = 500, groups = 4)

plot(df)

set.seed(123)

kmi <- kmeans_iterations(df, centers = 3)

plot(kmi)

}
