% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\name{mnist_plot_digits}
\alias{mnist_plot_digits}
\title{Plot some digits from train mnist data}
\usage{
mnist_plot_digits(ids = NULL)
}
\arguments{
\item{ids}{Rows to show.}
}
\description{
Plot some digits from train mnist data
}
\examples{

mnist_plot_digits(1)

mnist_plot_digits(c(10, 20, 40, 22))

set.seed(123)

mnist_plot_digits(sample(seq(60000), size = 16))

}
