% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnist_train}
\alias{mnist_train}
\title{MNIST train data}
\format{
A data frame with 60000 rows and 785 variables.
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_train
}
\description{
\itemize{
\item Original source    http://yann.lecun.com/exdb/mnist/
\item Converted to csv   https://pjreddie.com/projects/mnist-in-csv/
\item Downloaded from    https://www.kaggle.com/datasets/oddrationale/mnist-in-csv
}
}
\keyword{datasets}
