% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{sim_group_for_clusters}
\alias{sim_group_for_clusters}
\title{Generate data sets to apply a clustering algorithm}
\usage{
sim_group_for_clusters(n = 1000, groups = 3, props = NULL)
}
\arguments{
\item{n}{An intenger}

\item{groups}{An integer}

\item{props}{A vector of probabilties with length \code{groups}.}
}
\description{
Generate data sets to apply a clustering algorithm
}
\examples{

set.seed(123456)

df <- sim_group_for_clusters()

plot(df)

}
