% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{sim_groups}
\alias{sim_groups}
\title{Generate data sets to apply clustering algorithms}
\usage{
sim_groups(n = 1000, groups = 3, props = NULL)
}
\arguments{
\item{n}{An integer.}

\item{groups}{An integer}

\item{props}{A vector of probabilities with length \code{groups}.}
}
\description{
Generate data sets to apply clustering algorithms
}
\examples{

set.seed(123456)

df <- sim_groups()

plot(df)

}
