% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_2}
\alias{sim_quasianscombe_set_2}
\title{Generate \emph{quasi} Anscombe data sets Type 2}
\usage{
sim_quasianscombe_set_2(
  df,
  fun = function(x) {     x^2 },
  residual_factor = 0.25
)
}
\arguments{
\item{df}{A data frame from \code{sim_quasianscombe_set_1} (or similar).}

\item{fun}{A function to apply, this is applied to normalized version of \code{x}.}

\item{residual_factor}{Numeric value to multiply residual to modify their
variance.}
}
\description{
Data sets \emph{Type 2} shows how a no linear realtionship between \code{x} and \code{y} can
lead in the same regression model (in terms of parameter values) of
the \emph{Type 1}.
}
\examples{

df <- sim_quasianscombe_set_1()

dataset2 <- sim_quasianscombe_set_2(df)

dataset2

plot(dataset2)

plot(sim_quasianscombe_set_2(df, residual_factor = 0))

fun1 <- function(x){ sin(x*diff(range(x))) }

plot(sim_quasianscombe_set_2(df, fun = fun1))

fun2 <- abs

plot(sim_quasianscombe_set_2(df, fun = fun2))

fun3 <- function(x){ (x - mean(x)) * sin(x*diff(range(x))) }

plot(sim_quasianscombe_set_2(df, fun = fun3))

}
