% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_3}
\alias{sim_quasianscombe_set_3}
\title{Generate \emph{quasi} Anscombe data sets Type 3}
\usage{
sim_quasianscombe_set_3(
  df,
  prop = 0.05,
  beta1_factor = 0.5,
  residual_factor = 0.25
)
}
\arguments{
\item{df}{A data frame from \code{sim_quasianscombe_set_1} (or similar).}

\item{prop}{The proportion of value to modify as outliers.}

\item{beta1_factor}{Numeric value to modify the beta1 value.}

\item{residual_factor}{Numeric value to multiply residual to modify their
variance.}
}
\description{
Data sets \emph{Type 3} get some outliers but conserving the $x$ mean and
the same coefficients -but different significance- of the adjusted linear
model.
}
\details{
This function will:
\itemize{
\item Calculate the linear regression model and will calculate new trend using
0.5 times beta1
\item Take \code{prop}\% values from the greater \code{2*prop} \code{x} values and modify the
related \code{y} value to get the original estimation of \code{beta1}
\item Apply \code{residual_factor} factor to residual to get minor variance and
better visual impresion of the outlier effect.
}
}
\examples{

df <- sim_quasianscombe_set_1()

dataset3 <- sim_quasianscombe_set_3(df)

dataset3

plot(dataset3)

plot(sim_quasianscombe_set_3(df, prop = 0.1, residual_factor = 0))

}
