% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_4}
\alias{sim_quasianscombe_set_4}
\title{Generate \emph{quasi} Anscombe data sets Type 4}
\usage{
sim_quasianscombe_set_4(df, rescale_to = c(0.1, 0.2), prop = 0.15)
}
\arguments{
\item{df}{A data frame from \code{sim_quasianscombe_set_1} (or similar).}

\item{rescale_to}{Rescale the x value before create the second cluster.}

\item{prop}{The proportion of value to modify as the second group/cluster.}
}
\description{
Data sets \emph{Type 4} recreate two cluster keeping the coefficient of the original
regression model.
}
\details{
This function will:
\itemize{
\item Disorder the order of \code{x} values.
\item Rescale the \code{x} value to specific original quantiles.
\item Then take a porportion of value and traslate to left keeppping the original
mean of \code{x}.
\item Finally add some value to the associated \code{y} value and substract to the
complement group to have the same regression model in term of coefficients.
}
}
\examples{

df <- sim_quasianscombe_set_1()

dataset4 <- sim_quasianscombe_set_4(df)

dataset4

# plot(df)

plot(dataset4)

plot(sim_quasianscombe_set_4(df, rescale_to = c(0, .1), prop = 0.5))


}
