% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_6}
\alias{sim_quasianscombe_set_6}
\title{Generate \emph{quasi} Anscombe data sets Type 6: Simpson's Paradox}
\usage{
sim_quasianscombe_set_6(df, groups = 3, b1_factor = -1, residual_factor = 0.25)
}
\arguments{
\item{df}{A data frame from \code{sim_quasianscombe_set_1} (or similar).}

\item{groups}{Number of groups to separate \code{x} values.}

\item{b1_factor}{A numeric value get the slope in each group from $beta_1$.}

\item{residual_factor}{Numeric value to multiply residual to modify their
variance.}
}
\description{
Data sets \emph{Type 6} recreates the phenomenon of Simpon's paradox.
}
\details{
This function will take \code{x} vector and separate \code{groups} groups to apply
a local model with a modified regression using the \code{b1_factor} factor.

The residual will be multiply with a value between 0 and 1 to make the visual
effect greater.
}
\examples{

df <- sim_quasianscombe_set_1()

dataset6 <- sim_quasianscombe_set_6(df)

dataset6

plot(dataset6)

}
