% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_xy.R
\name{sim_response_xy}
\alias{sim_response_xy}
\title{Generate data sets to apply binary classifiers}
\usage{
sim_response_xy(
  n = 500,
  x_dist = purrr::partial(runif, min = -1, max = 1),
  y_dist = x_dist,
  relationship = function(x, y) x > y,
  noise = 0.2
)
}
\arguments{
\item{n}{An intenger}

\item{x_dist}{A random number generation function.}

\item{y_dist}{A random number generation function.}

\item{relationship}{A function specify the relationship between x, y and
the response. A function f(x, y) need return a logical value.}

\item{noise}{A number between 0 and 1.}
}
\description{
Generate data sets to apply binary classifiers
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 500)

df

plot(df)

}
