% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{sim_xy}
\alias{sim_xy}
\title{Generate data sets to apply regression methods}
\usage{
sim_xy(
  n = 500,
  beta0 = 3,
  beta1 = 0.5,
  x_dist = purrr::partial(rnorm, mean = 5, sd = 1),
  error_dist = purrr::partial(rnorm, sd = 0.5)
)
}
\arguments{
\item{n}{Number of observations}

\item{beta0}{beta0, default value: 3,}

\item{beta1}{beta1, default value: 0.5}

\item{x_dist}{A random number generation function. Default is a \code{rnorm}
with mean 5 and sd 1.}

\item{error_dist}{A random number generation function. Default is a \code{rnorm}
with mean 0 and sd 0.5.}
}
\description{
Generate data sets to apply regression methods
}
\examples{

df <- sim_xy()

df

plot(df)

klassets:::plot.klassets_xy(setNames(cars, c("x", "y")))

}
