% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{apply_kmeans_clust}
\alias{apply_kmeans_clust}
\title{Apply K-means to \code{klassets_cluster} object}
\usage{
apply_kmeans_clust(df, centers = 3, ...)
}
\arguments{
\item{df}{A \code{klassets_cluster} object.}

\item{centers}{A numeric value to pass to \code{kmeans_iterations} function
The famous k parameter.}

\item{...}{Extra parameters for \code{kmeans_iterations} function.}
}
\description{
Apply K-means to \code{klassets_cluster} object
}
\examples{

set.seed(12)

df <- sim_groups_for_clusters(n = 200, groups = 3)

plot(df)

set.seed(124)

dfc <- apply_kmeans_clust(df, centers = 4, max_iterations = 6)

plot(dfc)

}
