% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_response.R
\name{apply_logistic_regression}
\alias{apply_logistic_regression}
\title{Apply Logistic regression to \code{klassets_response_xy} object}
\usage{
apply_logistic_regression(df, order = 1, stepwise = FALSE, verbose = FALSE)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{order}{Order of values of x and y.}

\item{stepwise}{A logical value to indicate to perform stepwise.}
}
\description{
Apply Logistic regression to \code{klassets_response_xy} object
}
\examples{

set.seed(123)

df <- sim_response_xy(n = 500, relationship = function(x, y) x**2 > y)

plot(df)

class(df)

df <- apply_logistic_regression(df)

class(df)

plot(df)


}
