% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_regression_random_forest}
\alias{fit_regression_random_forest}
\title{Fit regression random forest to \code{klassets_xy} object}
\usage{
fit_regression_random_forest(
  df,
  ntree = 500L,
  maxdepth = Inf,
  alpha = 0.05,
  trace = TRUE,
  ...
)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{ntree}{Number of trees to grow for the forest.}

\item{maxdepth}{Max depth of the tree. Same used in \code{partykit::ctree_control}.}

\item{alpha}{Alpha value, same used in \code{partykit::ctree_control}}

\item{trace}{A logical indicating if a progress bar shall be printed while the forest grows.}

\item{...}{Options for \code{partykit::ctree_control}.}
}
\description{
Fit regression random forest to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dflm <- fit_regression_random_forest(df)

dflm

plot(dflm)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 3 * sin(x) + 5 * sqrt(abs(x)))

plot(df)

plot(fit_regression_random_forest(df))

# default
plot(fit_regression_random_forest(df))

}
