% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\name{fit_regression_tree}
\alias{fit_regression_tree}
\title{Fit regression tree to \code{klassets_xy} object}
\usage{
fit_regression_tree(df, maxdepth = Inf, alpha = 0.05, ...)
}
\arguments{
\item{df}{A object from \code{sim_response_xy}.}

\item{maxdepth}{Max depth of the tree. Same used in \code{partykit::ctree_control}.}

\item{alpha}{Alpha value, same used in \code{partykit::ctree_control}}

\item{...}{Options for \code{partykit::ctree_control}.}
}
\description{
Fit regression tree to \code{klassets_xy} object
}
\examples{

df <- sim_xy()

df

dflm <- fit_regression_tree(df)

dflm

plot(dflm)

df <- sim_xy(1000)
df <- dplyr::mutate(df, y = y + 10 * sin(x) + sqrt(abs(x)))

plot(df)

plot(fit_regression_tree(df, maxdepth = 3))

# default
plot(fit_regression_tree(df))

}
