% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_1}
\alias{sim_quasianscombe_set_1}
\title{Generate \emph{quasi} Anscombe data sets Type 1}
\usage{
sim_quasianscombe_set_1(
  n = 100,
  beta0 = 3,
  beta1 = 0.5,
  error_sd = 0.5,
  x_mean = 5,
  x_sd = 1,
  x_dist = "norm",
  seed = 123
)
}
\arguments{
\item{n}{n, default value: 100}

\item{beta0}{beta0, default value: 3,}

\item{beta1}{beta1, default value: 0.5}

\item{error_sd}{error_sd, default value: 0.1}

\item{x_mean}{x_mean, default value: 5}

\item{x_sd}{x_sd, default value: 1}

\item{x_dist}{x_dist, default value: "norm"}

\item{seed}{seed, default value: 1234}
}
\description{
This function generate a data set \emph{Type 1}  creating first a \code{x} a random vector
then apply a linear transformation using \code{beta0} and \code{beta1} and finally
adding a normal distributed noise using \code{error_sd} creating \code{y} values.
}
\details{
This is the \emph{typical} example when regression analysis is taught.
}
\examples{

df <- sim_quasianscombe_set_1()

df

plot(df)

plot(sim_quasianscombe_set_1(x_sd = 0))

plot(sim_quasianscombe_set_1(error_sd = 0))

plot(sim_quasianscombe_set_1(x_mean = 0, x_sd = 0))

}
