% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasianscombe.R
\name{sim_quasianscombe_set_5}
\alias{sim_quasianscombe_set_5}
\title{Generate \emph{quasi} Anscombe data sets Type 5}
\usage{
sim_quasianscombe_set_5(df, fun = identity, residual_factor = 10)
}
\arguments{
\item{df}{A data frame from \code{sim_quasianscombe_set_1} (or similar).}

\item{fun}{A function to apply to the index to multiply the residuals of
the original model.}

\item{residual_factor}{Numeric value to multiply residual to modify their
variance.}
}
\description{
Data sets \emph{Type 5} recreates the phenomenon of heteroskedasticity in
the residuals.
}
\details{
This function will take residuals $e_i$ and then get $e'_i = e_i * fun(i)$
and then rescale the $e'_i$ to the range of $e_i$.
}
\examples{

df <- sim_quasianscombe_set_1()

dataset5 <- sim_quasianscombe_set_5(df)

dataset5

plot(dataset5)

plot(sim_quasianscombe_set_5(df, fun = rev))

plot(sim_quasianscombe_set_5(df, fun = sqrt))

plot(sim_quasianscombe_set_5(df, fun = log))

plot(sim_quasianscombe_set_5(df, fun = function(x) x^(1+0.6)))

}
